refactor: arange the image file and rename all image and arange component on document
