+++ src/components/Apropos/apropos.tsx
@@ -1,15 +1,11 @@
-import Image from '../assets/Asja1.jpg';
-import Image2 from '../assets/Asja2.jpg';
-import FitbitIcon from '@mui/icons-material/Fitbit';
-import PinIcon from '@mui/icons-material/Pin';
-import Face2Icon from '@mui/icons-material/Face2';
-import AnchorIcon from '@mui/icons-material/Anchor';
-export default function home2() {
+import master from '../../assets/master_2025.jpg';
+
+export default function Apropos() {
   return (
     <>
         <div className='flex flex-wrap justify-center items-center w-full h-max bg-teal-800 text-white z-10'>
           <div className="flex-col w-1/2 px-15 py-15">
-            <img src={Image} className='flex-1/2 p-10'/>
+            <img src={master} className='flex-1/2 p-10'/>
             <h1 className='flex text-4xl px-10 pt-0  font-bold text-white'>Notre Mission</h1>
             <div className='px-10 text-lg pt-0  '>
               <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
@@ -26,14 +22,7 @@ export default function home2() {
               Lorem ipsum dolor sit amet, consectetur adipiscing elit, 
               sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
             </p>
-            <img src={Image2} alt="" className='flex-1/2 p-10'/>
-          </div>
-          <div className='flex m-5 w-full pb-15 px-15 justify-center items-center'>
-            <p className='mx-20'><FitbitIcon /> logo impsum</p>
-            <p className='mx-20'><PinIcon /> logo impsum</p>
-            <p className='mx-20'><Face2Icon /> logo impsum</p>
-            <p className='mx-20'><AnchorIcon /> logo impsum</p>
-
+            <img  alt="" className='flex-1/2 p-10'/> /** mbola asina sary */
           </div>
         </div>
     </>
